style: Organized `*.cs` files to easily distinguish what `ValueResult` and `VoidResult`
