+++ tests/PowerUtils.Results.Tests/ValueResultTests.cs
@@ -6,13 +6,13 @@ using Xunit;
 
 namespace PowerUtils.Results.Tests
 {
-    public class ResultTests
+    public class ValueResultTests
     {
         [Fact]
         public void WithoutErrors_GetErrors_InvalidOperationException()
         {
             // Arrange
-            var result = Result.Ok();
+            var result = Result<FakeModel>.Ok(new());
 
 
             // Act
@@ -25,14 +25,30 @@ namespace PowerUtils.Results.Tests
         }
 
         [Fact]
-        public void EmptyErrorList_CreateVoidResult_IsErrorFalse()
+        public void WithErrors_GetValue_InvalidOperationException()
+        {
+            // Arrange
+            Result<FakeModel> result = Error.Failure("fake", "fake", "fake");
+
+
+            // Act
+            var act = Record.Exception(() => result.Value);
+
+
+            // Assert
+            act.Should()
+                .BeOfType<InvalidOperationException>();
+        }
+
+        [Fact]
+        public void EmptyErrorList_CreateValueResult_IsErrorFalse()
         {
             // Arrange
             var errors = new List<Error>();
 
 
             // Act
-            Result act = errors;
+            Result<FakeModel> act = errors;
 
 
             // Assert
@@ -41,14 +57,14 @@ namespace PowerUtils.Results.Tests
         }
 
         [Fact]
-        public void EmptyIErrorList_CreateVoidResult_IsErrorFalse()
+        public void EmptyIErrorList_CreateValueResult_IsErrorFalse()
         {
             // Arrange
             var errors = new List<IError>();
 
 
             // Act
-            Result act = errors;
+            Result<FakeModel> act = errors;
 
 
             // Assert
@@ -57,14 +73,14 @@ namespace PowerUtils.Results.Tests
         }
 
         [Fact]
-        public void EmptyErrorArray_CreateVoidResult_IsErrorFalse()
+        public void EmptyErrorArray_CreateValueResult_IsErrorFalse()
         {
             // Arrange
             var errors = Array.Empty<Error>();
 
 
             // Act
-            Result act = errors;
+            Result<FakeModel> act = errors;
 
 
             // Assert
@@ -73,14 +89,14 @@ namespace PowerUtils.Results.Tests
         }
 
         [Fact]
-        public void EmptyIErrorArray_CreateVoidResult_IsErrorFalse()
+        public void EmptyIErrorArray_CreateValueResult_IsErrorFalse()
         {
             // Arrange
             var errors = Array.Empty<IError>();
 
 
             // Act
-            Result act = errors;
+            Result<FakeModel> act = errors;
 
 
             // Assert
@@ -89,14 +105,14 @@ namespace PowerUtils.Results.Tests
         }
 
         [Fact]
-        public void NullErrorList_CreateVoidResult_IsErrorFalse()
+        public void NullErrorList_CreateValueResult_IsErrorFalse()
         {
             // Arrange
             List<Error> errors = null;
 
 
             // Act
-            Result act = errors;
+            Result<FakeModel> act = errors;
 
 
             // Assert
@@ -105,14 +121,14 @@ namespace PowerUtils.Results.Tests
         }
 
         [Fact]
-        public void NullIErrorList_CreateVoidResult_IsErrorFalse()
+        public void NullIErrorList_CreateValueResult_IsErrorFalse()
         {
             // Arrange
             List<IError> errors = null;
 
 
             // Act
-            Result act = errors;
+            Result<FakeModel> act = errors;
 
 
             // Assert
@@ -121,14 +137,14 @@ namespace PowerUtils.Results.Tests
         }
 
         [Fact]
-        public void NullErrorArray_CreateVoidResult_IsErrorFalse()
+        public void NullErrorArray_CreateValueResult_IsErrorFalse()
         {
             // Arrange
             Error[] errors = null;
 
 
             // Act
-            Result act = errors;
+            Result<FakeModel> act = errors;
 
 
             // Assert
@@ -137,19 +153,61 @@ namespace PowerUtils.Results.Tests
         }
 
         [Fact]
-        public void NullIErrorArray_CreateVoidResult_IsErrorFalse()
+        public void NullIErrorArray_CreateValueResult_IsErrorFalse()
         {
             // Arrange
             IError[] errors = null;
 
 
             // Act
-            Result act = errors;
+            Result<FakeModel> act = errors;
 
 
             // Assert
             act.IsError.Should()
                 .BeFalse();
+        }
+
+        [Fact]
+        public void WithValue_GetValue_SameValue()
+        {
+            // Arrange
+            var id = 343;
+            var name = "FakeName";
+
+            var result = Result<FakeModel>.Ok(new(id, name));
+
+
+            // Act
+            var act = result.Value;
+
+
+            // Assert
+            act.Id.Should()
+                .Be(id);
+            act.Name.Should()
+                .Be(name);
+        }
+
+        [Fact]
+        public void Value_ImplicitAssignment_SameValue()
+        {
+            // Arrange
+            var id = 343;
+            var name = "FakeName";
+
+            var model = new FakeModel(id, name);
+
+
+            // Act
+            Result<FakeModel> act = model;
+
+
+            // Assert
+            act.Value.Id.Should()
+                .Be(id);
+            act.Value.Name.Should()
+                .Be(name);
         }
 
         [Fact]
@@ -172,7 +230,7 @@ namespace PowerUtils.Results.Tests
 
 
             // Act
-            Result act = errors;
+            Result<FakeModel> act = errors;
 
 
             // Assert
@@ -218,7 +276,7 @@ namespace PowerUtils.Results.Tests
 
 
             // Act
-            Result act = errors;
+            Result<FakeModel> act = errors;
 
 
             // Assert
@@ -228,7 +286,7 @@ namespace PowerUtils.Results.Tests
             act.Errors.Should()
                 .ContainSingle(s =>
                     s.Property == property1
-                    &&
+                &&
                     s.Code == code1
                     &&
                     s.Description == description1
@@ -237,7 +295,7 @@ namespace PowerUtils.Results.Tests
             act.Errors.Should()
                 .ContainSingle(s =>
                     s.Property == property2
-                    &&
+                &&
                     s.Code == code2
                     &&
                     s.Description == description2
@@ -262,7 +320,7 @@ namespace PowerUtils.Results.Tests
 
 
             // Act
-            Result act = new Error(property1, code1, description1);
+            Result<FakeModel> act = new Error(property1, code1, description1);
             act.AddError(new Error(property2, code2, description2));
             act.AddError(new Error(property3, code3, description3));
 
@@ -313,7 +371,7 @@ namespace PowerUtils.Results.Tests
 
 
             // Act
-            var act = new Result();
+            var act = Result<FakeModel>.Ok(null);
             act.AddError(new Error(property1, code1, description1));
             act.AddError(property2, code2, description2);
 
@@ -342,10 +400,11 @@ namespace PowerUtils.Results.Tests
         }
 
         [Fact]
-        public void WithoutError_GetType_Success()
+        public void WithoutError_GetType_TypeOfModel()
         {
             // Arrange
-            var result = Result.Ok();
+            var model = new FakeModel();
+            var result = Result<FakeModel>.Ok(model);
 
 
             // Act
@@ -353,14 +412,14 @@ namespace PowerUtils.Results.Tests
 
 
             // Assert
-            act.Should().Be(typeof(Success));
+            act.Should().Be(typeof(FakeModel));
         }
 
         [Fact]
         public void WithConflictError_GetType_ConflictError()
         {
             // Arrange
-            Result result = Error.Conflict("prop", "code", "disc");
+            Result<FakeModel> result = Error.Conflict("prop", "code", "disc");
 
 
             // Act
@@ -375,7 +434,7 @@ namespace PowerUtils.Results.Tests
         public void WithNotFoundError_GetType_NotFoundError()
         {
             // Arrange
-            Result result = Error.NotFound("prop", "code", "disc");
+            Result<FakeModel> result = Error.NotFound("prop", "code", "disc");
 
 
             // Act
@@ -387,44 +446,38 @@ namespace PowerUtils.Results.Tests
         }
 
         [Fact]
-        public void CustomError_ResultFrom_Result()
+        public void ResultWithValue_ImplicitAssignment_Value()
         {
             // Arrange
-            var property = "fakeCustomProperty";
-            var code = "fakeCustomCode";
-            var description = "fakeCustomDescription";
+            var id = 45615;
+            var name = "fake name";
 
-            var error = new CustomError(property, code, description);
+            Result<FakeModel> result = new FakeModel { Id = id, Name = name };
 
 
             // Act
-            var act = Result.From(error);
+            FakeModel act = result;
 
 
             // Assert
-            act.FirstError().Property.Should()
-                .Be(property);
-            act.FirstError().Code.Should()
-                .Be(code);
-            act.FirstError().Description.Should()
-                .Be(description);
-
-            act.GetType().Should().Be(typeof(CustomError));
+            act.Id.Should().Be(id);
+            act.Name.Should().Be(name);
         }
 
 
+
         [Fact]
-        public void ValueResultWithErrors_ImplicitAssignment_VoidResult()
+        public void VoidResultWithErrors_ImplicitAssignment_ValueResult()
         {
             // Arrange
             var property = "fakeProperty";
             var code = "fakeCode";
             var description = "fakeDescription";
-            Result<FakeModel> result = Error.Conflict(property, code, description);
+            Result result = Error.Conflict(property, code, description);
 
 
             // Act
-            Result act = result;
+            Result<FakeModel> act = result;
 
 
             // Assert
@@ -439,36 +492,40 @@ namespace PowerUtils.Results.Tests
         }
 
         [Fact]
-        public void ValueResultWithoutErrors_ImplicitAssignment_VoidResult()
+        public void VoidResultWithoutErrors_ImplicitAssignment_ValueResult()
         {
             // Arrange
-            Result<FakeModel> result = Array.Empty<Error>();
+            Result result = Array.Empty<Error>();
 
 
             // Act
-            Result act = result;
+            Result<FakeModel> act = result;
 
 
             // Assert
             act.IsError.Should()
                 .BeFalse();
+            act.Value.Should()
+                .BeNull();
         }
 
 #if NET5_0
         [Fact]
-        public void ValueResultNull_ImplicitAssignment_VoidResult()
+        public void VoidResultNull_ImplicitAssignment_ValueResult()
         {
             // Arrange
-            Result<FakeModel> result = null;
+            Result result = null;
 
 
             // Act
-            Result act = result;
+            Result<FakeModel> act = result;
 
 
             // Assert
             act?.IsError.Should()
                 .BeFalse();
+            act.Value.Should()
+                .BeNull();
         }
 #endif
     }
